requirements.txt
pysam>=0.20.0
biopython>=1.74
numpy>=1.24.3
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.1
statsmodels>=0.13.2
Unidecode>=1.3.6